fix: 그룹조회 반환값에 참가 멤버 수 추가
