+++ src/main/java/com/gloddy/server/apply/service/ApplyService.java
@@ -1,0 +1,13 @@
+package com.gloddy.server.apply.service;
+
+import com.gloddy.server.apply.repository.ApplyJpaRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class ApplyService {
+
+    private final ApplyJpaRepository applyJpaRepository;
+}
