+++ src/main/java/com/gloddy/server/group/dto/GroupResponse.java
@@ -4,14 +4,35 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
+import java.util.List;
+
 public class GroupResponse {
 
     @Getter
     @AllArgsConstructor
     @NoArgsConstructor
     public static class Create {
-
         private Long groupId;
+    }
+
+    @Getter
+    @AllArgsConstructor
+    public static class GetGroups {
+        List<GroupResponse.GetGroup> groups;
+    }
+
+    // TODO: 참가 멤버 수, 멤버 프로필
+    @Getter
+    @AllArgsConstructor
+    public static class GetGroup {
+        // String imageUrl;
+        String title;
+        String content;
+        // List<String> memberProfiles;
+        int memberCount;
+        String place;
+        LocalDate meetDate;
     }
 
 }
